+++ BxTextFieldUITests/BxTextFieldUILeftToRightTests.swift
@@ -11,6 +11,7 @@ import XCTest
 class BxTextFieldUILeftToRightTests: BxUITestCase {
     
     var textField : XCUIElement!
+    var rewritingSwitch : XCUIElement!
     
     override func setUp()
     {
@@ -20,6 +21,7 @@ class BxTextFieldUILeftToRightTests: BxUITestCase {
         let tablesQuery = app.tables
         
         textField = tablesQuery.children(matching: .cell).element(boundBy: 2).children(matching: .textField).element
+        rewritingSwitch = tablesQuery.children(matching: .cell).element(boundBy: 2).children(matching: .switch).element
         
         textField.tap()
         textField.typeText("0000000000")
@@ -111,10 +113,38 @@ class BxTextFieldUILeftToRightTests: BxUITestCase {
         coordinate.withOffset(CGVector(dx:-43, dy:0)).tap()
         textField.doubleTap()
         coordinate.withOffset(CGVector(dx:-48, dy:0)).press(forDuration: 1, thenDragTo: coordinate.withOffset(CGVector(dx:-78, dy:0)))
-        textField.typeText("12345")
         
+        textField.typeText("123")
+        XCTAssertEqual(textField.value as! String, "+0 (987) 123 - 10")
+        
+        textField.typeText("45")
         XCTAssertEqual(textField.value as! String, "+0 (987) 123 - 45 - 10")
         
+    }
+    
+    func testRewriting()
+    {
+        textField.typeText("09876543210")
+        XCTAssertEqual(textField.value as! String, "+0 (987) 654 - 32 - 10")
+        
+        let coordinate = textField.coordinate(withNormalizedOffset: CGVector(dx:1, dy:0))
+        coordinate.withOffset(CGVector(dx:-40, dy:0)).tap()
+        
+        textField.typeText("123")
+        XCTAssertEqual(textField.value as! String, "+0 (987) 654 - 32 - 10")
+        
+        textField.typeText("\u{8}\u{8}")
+        rewritingSwitch.tap()
+        textField.typeText("12345")
+        XCTAssertEqual(textField.value as! String, "+0 (987) 654 - 12 - 34")
+        
+        rewritingSwitch.tap()
+        coordinate.withOffset(CGVector(dx:-40, dy:0)).tap()
+        textField.typeText("987")
+        XCTAssertEqual(textField.value as! String, "+0 (987) 654 - 12 - 34")
+        textField.typeText("\u{8}\u{8}")
+        textField.typeText("987")
+        XCTAssertEqual(textField.value as! String, "+0 (987) 654 - 98 - 34")
     }
     
 }